If the needed template is not in the database, search it as a file in the directory "templates".
Posting new thread and answers has been implemented.

Some work on the navigation toolbars. Little fix to the sessions handling. Not if there is 
active session, it will be reused. This way, login from another browser does not logout the other browser.

Some more work have to be done on the cookie handling and navigation.