Implemented an option to select the embeded or external .css to be used. The CSS files are now processed by the template render and can contains template commands.
The external .css files improve the browser caching and the embeded .css improve rendering time of the web page, removing the flickering.

There is a massive file change, but the changes are small - only added .tpl suffix to the StrCatTemplate calls and some .less files changed to contain template commands.