Now the user related events are sent one by one only on change. The whole users set is sent only on the initialization of the chat.
The event handling service seems to work pretty consistently. Tests about memory leaks are needed. The flagSelf was removed from the users events and replaced with sending the first 8 symbols from the event session ID. This way, the clients can detect the
self events by comparing with their own cookie. 

The js code of the chat was refactored according to the new events and agressive automatic reconnection has been implemented.